fix: Adds k8s-netperf to the container image
This commit adds a go build to the Containerfile via multi-stage build. For this to work, also the Makefile was change to load the full repo into the container build context.

This also adds a simple example to the docs section on how to use the container image to performe the tests.